+++ CHANGELOG.md
@@ -40,6 +40,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ls *.jpg | machconvert --resize-x 800
 - ls *.jpg | machconvert --resize-y 800
 
+- ls *.jpg | machreduce -o result.zip --orientation landscape --nb-by-pages 4
+
 - ls *.jpg | machreduce -o result.odt
 - ls *.jpg | machreduce -o result.odp